+++ src/commands/index.ts
@@ -3,3 +3,4 @@ export * from './listAllImports';
 export * from './packageJsonToMarkdown';
 export * from './projectDirectoryStructure';
 export * from './showComponentHierarchy';
+export * from './componentHierarchyMarkdown';