+++ src/commands/packageJsonToMarkdown.ts
@@ -1,8 +1,7 @@
 import * as fs from 'fs';
 import * as path from 'path';
 
-import { Config } from '../config';
-import { FileSystemUtils } from '../filesystemUtils';
+import { Config, FileSystemUtils } from '@src';
 
 const fetch = require('node-fetch');