+++ src/index.ts
@@ -1,0 +1,3 @@
+export * from './componentManager';
+export * from './config';
+export * from './filesystemUtils';