+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 
 ## Version 1.1.0
 
+- New feature: You can now generate a markdown(Mermaid) file with the component hierarchy.
 - Bugfix: The List imports command now also find the require(...) packages.
+- Bugfix: ShowComponentHierarchy.html is no longer written to the workspace root.
 
 ## Version 1.0.1
 
-- Bugfix: The Show component hierarchy graph command did not work 
+- Bugfix: The Show component hierarchy graph command did not work
 
 ## Version 1.0.0